add security for update && delete goals

+++ backend/controllers/goalControllers.js
@@ -1,10 +1,10 @@
 const asyncHandler = require("express-async-handler")
 const Goal = require('../models/goalModel');
-
+const User = require('../models/userModel')
 // @desc Get goals
 // @route Get /api/goals
 const getGoals = asyncHandler(async (req,res) => {
-    const goal = await Goal.find()
+    const goal = await Goal.find({ user: req.user.id})
 
     res.status(200).json(goal);
 })
@@ -17,7 +17,8 @@ const setGoal = asyncHandler(async (req,res) => {
         throw new Error('Please add a text field')
     }
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
     res.status(200).json(goal)
 })
@@ -25,11 +26,26 @@ const setGoal = asyncHandler(async (req,res) => {
 // @desc UPDATE goal
 // @route UPDATE /api/goals/:id
 const updateGoal = asyncHandler(async (req,res) => {
+    const user = await User.findById(req.user.id)
     const goal = await Goal.findById(req.params.id)
 
     if(!goal){
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    // check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+    
+    // Make sure the looged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
+    }else{
+        console.log("ok")
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -42,9 +58,28 @@ const updateGoal = asyncHandler(async (req,res) => {
 // @desc DELETE goal
 // @route DELETE /api/goals/:id
 const deleteGoal = asyncHandler(async (req,res) => {
+    const user = await User.findById(req.user.id)
     const goal = await Goal.findById(req.params.id)
 
+    if(!goal){
+        res.status(400)
+        throw new Error('Goal not found')
+    }
+
+    // check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+    
+    // Make sure the looged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+    
     await goal.deleteOne()
+
     res.status(200).json({id: req.params.id})
 })
